moved control types from "trajectory_control.rs" to "control_types.rs" as the first step of disentangeling the monolith
